+++ packages/browser/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @glissade/browser
 
+## 0.79.0-pre.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/scene@0.79.0-pre.0
+  - @glissade/backend-canvas2d@0.79.0-pre.0
+  - @glissade/backend-dom@0.79.0-pre.0
+  - @glissade/element@0.79.0-pre.0
+  - @glissade/player@0.79.0-pre.0
+  - @glissade/core@0.79.0-pre.0
+
 ## 0.78.0
 
 ### Patch Changes